move the current frontend code to make theming easier
