+++ public/themes/default/css/editor.css
