+++ public/themes/default/pages/admin.html
@@ -5,8 +5,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>The Admin Page</title>
     <link rel="icon" type="image/png" href="/favicon.png" />
-    <link rel="stylesheet" href="/css/admin.css" async>
-    <script type="module" src="/js/admin.js" async defer></script>
+    <link rel="stylesheet" href="/themes/default/css/admin.css" async>
+    <script type="module" src="/themes/default/js/admin.js" async defer></script>
   </head>
   <body>
     <h1>Admin interface</h1>
