+++ public/themes/default/pages/editor.html
@@ -9,10 +9,10 @@
       type="module"
     ></script>
     <link rel="icon" type="image/png" href="/favicon.png" />
-    <link rel="stylesheet" href="/css/layout.css" />
-    <link rel="stylesheet" href="/css/editor.css" />
-    <link rel="stylesheet" href="/css/file-tree.css" />
-    <link rel="stylesheet" href="/css/cm6.css" />
+    <link rel="stylesheet" href="/themes/default/css/layout.css" />
+    <link rel="stylesheet" href="/themes/default/css/editor.css" />
+    <link rel="stylesheet" href="/themes/default/css/file-tree.css" />
+    <link rel="stylesheet" href="/themes/default/css/cm6.css" />
   </head>
   <body
     {% if projectMember %}data-project-member="true"{% endif %}
