+++ public/themes/default/pages/fragments/edit.html
@@ -1,4 +1,4 @@
-<link rel="stylesheet" href="/css/fragments/edit.css" />
+<link rel="stylesheet" href="/themes/default/css/fragments/edit.css" />
 
 <template id="project-edit-template">
   <dialog>
